+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # [unreleased]
 
 - ui: Basic recovery from failed attempts to use devices (e.g. flaky bluetooth)
+- ui: Display restart reason in case of premature failure
 
 # 0.3.1 [2026-09-05]